add src/discord_harness directory for easy ensurance that the test files will be using the installed package and not the one in the repository root
We want each test to import discord_harness as an installed package, if we put discord_harness in the root of the repository, and also are supposed to install a package that makes it available it will potentially be confusing and error-prone. This is why a src directory has been added where it otherwise wouldn't be in various of our python projects.