add simple pyproject.toml with setuptools as the build system and explicitly make src/ the place to find packages
Some versions of setuptools will look in the repository root for packages, others in the src directory if present. I do not want to think about compatibility if we change setuptools versions in the future, and I am configuring this in explicitly.

Setting [tool.setuptools.packages.find] makes it not fill in default package information and break. So this also has the name and version filled in immediately too.