add start of "gateway ready" payload data making code with some simple tests
The discord API will send a ready event with a list of each guild the user is in, with the ID and a field stating it's unavailable. Each guild the user is in will have a GUILD_CREATE event sent later with the information for it.

Some basic data has been filled in to start, with some default values for some stuff not specified. This will be driven further by what is necessary to test the harness/bot integration.

The avatar field is mandatory for discord.py .