add start of harness with an initialize method that when called will eventually trigger on_ready
Initializing the client with the name of a created user should probably result in that users information coming into the client and a ready event being fired. Right now we care about the ready event. We'll also want to fire off the guilds that they are in later.

A simple test has been added to go through the workflow of creating a new user, and initializing the client with that user account. The behavior tested for now is that on_ready will be called.