add SystemState class with next_id method for the start of storing/creating entities.
We're going to have various entities that live in memory such as Users, Guilds, etc. We're going to manage them through this collection, and this will have some utilities for working with them such as getting the next id to assign.

The assignment logic is simple to start and within the responsibility of the class. We're not using snowflake ids yet or separating concerns here. What we want is something simple and testable, it goes up by one with each call.