introduce ClientEventCommunicator class as implementation detail for communicating event information + defaults to the discord client (for now, subject to change) and update harness to use it
Each piece of the harness is going to need to communicate information to the client such as guild being created, a member being banned, etc, and the code would be simpler if it didn't have to worry about how it is delivered to the client. This is adding a basic abstraction for it.