move information from backend to new "discord_harness.database.information" package
This information is meant to be pulled from the system state regardless to any application specific code, so it should go in the "database" code rather than the backend code. The backend code right now is basically the application service/use case layer, though this architecture is only loosely based on clean/others.